avoid probe rewriting of p->m in &(p->m)
Fix issue #1830.

After the rewrite, the code approximately becomes
  &(({type _val; bpf_probe_read(&_val, sizeof(_val), &(p->m)); _val)

Firstly the rewriting is really unnecessary, and secondly
the compilation will fail since the addressOf cannot take address
of the rvalue _val.

This patch intends to fix the problem by avoiding
the rewriting in the first place. The implementation
assumes no nested addressOf operations which should be
true in most cases if not all.

Signed-off-by: Yonghong Song <yhs@fb.com>